interferometer tests moved
